Added a device reference to DeviceHardware and put hardware classes in separate files
